Update Blinds_program.ino
Now completely working with hold and tap, AND showing up in Home assistant as a device (not just an entity).

Still need to decide if I'll be reading what home assistant sends by default (i.e. a message to a MQTT "set" topic), or changing the program to accept that and send what needs sent. Seems a waste to have all this specialized JSON parsing without using it though.